refactor: rename files - remove .stub
