+++ src/lib/invoice_object.js
