+++ src/lib/invoice_request_abstract.js
