+++ src/lib/invoice_request_cancel.js
@@ -1,4 +1,4 @@
-const Invoice_RUDCnP = require("./stub.invoice_request_abstract_RUDCP_super");
+const Invoice_RUDCnP = require("./invoice_request_abstract_RUDCP_super");
 const { shazam_integer } = require("./utilities/aaa_index");
 /** @class Invoice_Cancel
  * @param {string} id The invoice_id of the invoice you want
