+++ src/lib/invoice_request_create.js
@@ -1,4 +1,4 @@
-const Invoice_Request = require("./stub.invoice_request_abstract");
+const Invoice_Request = require("./invoice_request_abstract");
 const { nanoid } = require("nanoid");
 const { shazam_max_length } = require("./utilities/aaa_index");
