+++ src/lib/invoice_request_publish.js
@@ -1,4 +1,4 @@
-const Invoice_RUDCnP = require("./stub.invoice_request_abstract_RUDCP_super");
+const Invoice_RUDCnP = require("./invoice_request_abstract_RUDCP_super");
 const { nanoid } = require("nanoid");
 const { shazam_integer, shazam_max_length } = require("./utilities/aaa_index");
